displays warning prompt when attempting to overwrite subset gpkg (#133)
* displays warning prompt when user attempts to overwrite an existing subset gpkg

* map app prompts before potential overwrite

* clean up old code

* speed up upstream cat fetching

---------

+++ modules/data_processing/subset.py
@@ -2,6 +2,8 @@ import logging
 import os
 from pathlib import Path
 from typing import List, Union
+from rich.prompt import Prompt
+from rich.console import Console
 
 from data_processing.file_paths import file_paths
 from data_processing.gpkg_utils import (
@@ -14,7 +16,7 @@ from data_processing.gpkg_utils import (
 from data_processing.graph_utils import get_upstream_ids
 
 logger = logging.getLogger(__name__)
-
+console = Console()
 subset_tables = [
     "divides",
     "divide-attributes",  # requires divides
@@ -30,16 +32,30 @@ subset_tables = [
 
 
 def create_subset_gpkg(
-    ids: Union[List[str], str], hydrofabric: Path, output_gpkg_path: Path, is_vpu: bool = False
+    ids: Union[List[str], str], hydrofabric: Path, output_gpkg_path: Path, is_vpu: bool = False, override_gpkg: bool = False
 ):
     # ids is a list of nexus and wb ids, or a single vpu id
     if not isinstance(ids, list):
         ids = [ids]
     output_gpkg_path.parent.mkdir(parents=True, exist_ok=True)
 
-    if os.path.exists(output_gpkg_path):
-        os.remove(output_gpkg_path)
-
+    if not override_gpkg:
+        if os.path.exists(output_gpkg_path):
+            response = Prompt.ask(
+                f"Subset geopackage at {output_gpkg_path} already exists. Are you sure you want to overwrite it?",
+                default="n",
+                choices=["y", "n"],
+            )
+            if response == "y":
+                console.print(f"Removing {output_gpkg_path}...", style="yellow")
+                os.remove(output_gpkg_path)
+            else:
+                console.print("Exiting...", style="bold red")
+                exit()
+    else:
+        if os.path.exists(output_gpkg_path):
+            os.remove(output_gpkg_path)
+            
     create_empty_gpkg(output_gpkg_path)
     logger.info(f"Subsetting tables: {subset_tables}")
     for table in subset_tables:
@@ -55,8 +71,18 @@ def create_subset_gpkg(
 def subset_vpu(
     vpu_id: str, output_gpkg_path: Path, hydrofabric: Path = file_paths.conus_hydrofabric
 ):
-    if output_gpkg_path.exists():
-        os.remove(output_gpkg_path)
+    if os.path.exists(output_gpkg_path):
+        response = Prompt.ask(
+            f"Subset geopackage at {output_gpkg_path} already exists. Are you sure you want to overwrite it?",
+            default="n",
+            choices=["y", "n"],
+        )
+        if response == "y":
+            console.print(f"Removing {output_gpkg_path}...", style="yellow")
+            os.remove(output_gpkg_path)
+        else:
+            console.print("Exiting...", style="bold red")
+            exit()
 
     create_subset_gpkg(vpu_id, hydrofabric, output_gpkg_path=output_gpkg_path, is_vpu=True)
     logger.info(f"Subset complete for VPU {vpu_id}")
@@ -68,6 +94,7 @@ def subset(
     hydrofabric: Path = file_paths.conus_hydrofabric,
     output_gpkg_path: Path = Path(),
     include_outlet: bool = True,
+    override_gpkg: bool = False # only gets set to true through map app
 ):
     upstream_ids = list(get_upstream_ids(cat_ids, include_outlet))
 
@@ -78,6 +105,6 @@ def subset(
         paths = file_paths(output_folder_name)
         output_gpkg_path = paths.geopackage_path
 
-    create_subset_gpkg(upstream_ids, hydrofabric, output_gpkg_path)
+    create_subset_gpkg(upstream_ids, hydrofabric, output_gpkg_path, override_gpkg=override_gpkg)
     logger.info(f"Subset complete for {len(upstream_ids)} features (catchments + nexuses)")
     logger.debug(f"Subset complete for {upstream_ids} catchments")
